% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bed_usage.R
\name{bed_usage_ui}
\alias{bed_usage_ui}
\alias{bed_usage_server}
\title{Create a page for displaying bed usage}
\usage{
bed_usage_ui(id, cause)

bed_usage_server(id, board_ref, pin_name, selected_trusts)
}
\arguments{
\item{id}{Unique Shiny module ID to link the UI and Server components together.}

\item{cause}{Display name of the illness that a patient occupies a bed for. Write in lower case
unless capitalisation is needed to denote an acronym of proper noun, since this will be
displayed as text in the middle of sentences where needed.}

\item{board_ref}{A pointer to a Posit Connect board created by \link[pins:board_connect]{pins::board_connect} or an
alternative location.}

\item{pin_name}{Name of the pin on the Posit Connect board containing the bed usage data for this
particular illness. This data should have been pre-processed. If the data is in parquet format,
then the nanoparquet package is required.}

\item{selected_trusts}{Character vector of the trust ID codes that the user has selected in the
filter.}
}
\description{
Dashboard content for displaying summary information about bed-occupancy for a particular cause
of illness.
}
