% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bed_usage_plot.R
\name{bed_usage_plot_ui}
\alias{bed_usage_plot_ui}
\alias{bed_usage_plot_server}
\title{Add a time series plot of bed usage}
\usage{
bed_usage_plot_ui(id, cause)

bed_usage_plot_server(id, filtered_data, cause)
}
\arguments{
\item{id}{Unique Shiny module ID to link the UI and Server components together.}

\item{cause}{Display name of the illness that a patient occupies a bed for. Write in lower case
unless capitalisation is needed to denote an acronym of proper noun, since this will be
displayed as text in the middle of sentences where needed.}

\item{filtered_data}{A \link[shiny:reactiveValues]{shiny::reactiveValues} object containing an entry called \code{full} which has
the data only for the trusts selected in the user's filtering options.}
}
\description{
Shiny module for including a time series plot of bed usage.
}
