% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_over_time.R
\name{change_over_time}
\alias{change_over_time}
\title{Get change in bed usage since the start of a time period}
\usage{
change_over_time(.data, time_period)
}
\arguments{
\item{.data}{A tibble or data frame showing the bed usage at all trusts at all available points
in time.}

\item{time_period}{A time period to denote how far back in time from the latest reading we want
to go in order to see the change over time. This should be a period object, which can be
created using lubridate methods such as \code{\link[lubridate:period]{lubridate::years}}\code{(1)} to create a comparison with 1
year ago, or \code{\link[lubridate:period]{lubridate::weeks}}\code{(2)} to create a comparison with 2 weeks ago.}
}
\value{
A summary data frame containing the Trust \code{code}, the \code{bed_type}, and the \code{change}
column containing the difference between the starting value and the final value.
}
\description{
This creates a summary table of how bed usage has changed over the course of a time period.
The value reported is the total difference between the start of the time period and the latest
counts.

Where a negative value is returned as the \code{change} value, this denotes there was a \emph{decrease} in
the bed occupancy from the start of the time period to the latest date in the data.
}
