% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trust_list.R
\name{get_trust_list}
\alias{get_trust_list}
\title{Get a table of the trust names}
\usage{
get_trust_list(board_ref, pin_name)
}
\arguments{
\item{board_ref}{Connection to a Pins board.}

\item{pin_name}{Name of the pin that is on the Posit Connect board. The connection to the board
is created automatically through Environment Variable authentication, meaning that the
environment variables \code{CONNECT_API_KEY} and \code{CONNECT_SERVER} should be configured.

The format of the pin is assumed to be "parquet".}
}
\value{
A data frame containing a column of Trust Name (\code{name}), Region (\code{nhs_england_region}),
and Trust ID code (\code{code}).
}
\description{
Get a table of the trust names
}
