% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_year_end.R
\name{get_year_end}
\alias{get_year_end}
\title{Determine the year end for a date.}
\usage{
get_year_end(date)
}
\arguments{
\item{date}{A Date object}
}
\value{
An integer for the year when the current season ends. If the \code{date} value is in December,
it will return the following calendar year.
}
\description{
The year end is defined as the year where the current season ends. The start of a season is
defined as 1st December, therefore the end of a season is 30th November.
}
\examples{
get_year_end(Sys.Date())
}
