% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_data.R
\name{load_all_data}
\alias{load_all_data}
\title{Load data for all years from pins.}
\usage{
load_all_data(board_ref, starting_year_end, pin_name_template)
}
\arguments{
\item{board_ref}{A pointer to a Posit Connect board created by \code{\link[pins:board_connect]{pins::board_connect()}} or an
alternative location.}

\item{starting_year_end}{The (numeric) year end of the first year of data that should be
imported. e.g., if the first year of data covers the 2021-2022 season, then the year end is
2022.}

\item{pin_name_template}{Character template for the name of the pin containing the data. This
uses \code{\link[=sprintf]{sprintf()}} notation, so use \verb{\%d} as a placeholder for where the year should be entered.
If the data is in parquet format, then the nanoparquet package is required.}
}
\value{
A data frame containing the imported data and an additional column of year_end values.
}
\description{
Connect to a Posit Connect board and then pull data for all the years from the respective pins.
}
