% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_time_series.R
\name{plot_annual_time_series}
\alias{plot_annual_time_series}
\title{Create time series plot of data over multiple years}
\usage{
plot_annual_time_series(.data)
}
\arguments{
\item{.data}{A data frame containing the time series data.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} of the time series of each year.
}
\description{
Create time series plot of data over multiple years
}
