% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_time_series.R
\name{plot_region_time_series}
\alias{plot_region_time_series}
\alias{summarise_region_time_series}
\title{Summarise and plot time series data across all regions for this year}
\usage{
plot_region_time_series(.data, cause)

summarise_region_time_series(.data)
}
\arguments{
\item{.data}{A data frame containing the time series data.}

\item{cause}{Display name of the illness that a patient occupies a bed for. Write in lower case
unless capitalisation is needed to denote an acronym of proper noun, since this will be
displayed as text in the middle of sentences where needed.}
}
\value{
For \code{plot_region_time_series}, a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} of the total usage in each region.

From \code{summarise_region_time_series}, a data frame summarising the overall usage on each
day of eeach region in the latest year.
}
\description{
Summarise and plot time series data across all regions for this year
}
