% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_box_change.R
\name{value_box_change_ui}
\alias{value_box_change_ui}
\alias{value_box_change_server}
\title{Single value box for showing change over time}
\usage{
value_box_change_ui(id, period_label)

value_box_change_server(id, filtered_data, time_period)
}
\arguments{
\item{id}{Unique Shiny module ID to link the UI and Server components together.}

\item{period_label}{Character description of the time period that will be quoted in the value box title.
The title will be prefixed with the words "Change since", before printing your period label afterwards.
Aim to keep this concise.}

\item{filtered_data}{A \link[shiny:reactiveValues]{shiny::reactiveValues} object containing an entry called \code{full} which has
the data only for the trusts selected in the user's filtering options.}

\item{time_period}{A time period to denote how far back in time from the latest reading we want
to go in order to see the change over time. This should be a period object, which can be
created using lubridate methods such as \code{\link[lubridate:period]{lubridate::years}}\code{(1)} to create a comparison with 1
year ago, or \code{\link[lubridate:period]{lubridate::weeks}}\code{(2)} to create a comparison with 2 weeks ago.}
}
\description{
One value box for showing the change in bed occupancy over a given time period
}
