% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_value_box_row.R
\name{value_box_row_ui}
\alias{value_box_row_ui}
\alias{value_box_row_server}
\title{Add a row of key statistics in value boxes}
\usage{
value_box_row_ui(id, cause)

value_box_row_server(id, selected_trusts, filtered_data)
}
\arguments{
\item{id}{Unique Shiny module ID to link the UI and Server components together.}

\item{cause}{Display name of the illness that a patient occupies a bed for. Write in lower case
unless capitalisation is needed to denote an acronym of proper noun, since this will be
displayed as text in the middle of sentences where needed.}

\item{selected_trusts}{Character vector of the trust ID codes that the user has selected in the
filter.}

\item{filtered_data}{A \link[shiny:reactiveValues]{shiny::reactiveValues} object containing an entry called \code{full} which has
the data only for the trusts selected in the user's filtering options.}
}
\description{
Shiny module for including a row of 4 key statistics as \link[bslib:value_box]{bslib::value_box} objects.
}
