% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_time_series.R
\name{plot_annual_time_series}
\alias{plot_annual_time_series}
\alias{summarise_annual_time_series}
\title{Summarise and plot time series data over multiple years}
\usage{
plot_annual_time_series(.data, cause)

summarise_annual_time_series(.data)
}
\arguments{
\item{.data}{A data frame containing the time series data.}

\item{cause}{Display name of the illness that a patient occupies a bed for. Write in lower case
unless capitalisation is needed to denote an acronym of proper noun, since this will be
displayed as text in the middle of sentences where needed.}
}
\value{
For \code{plot_annual_time_series}, a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} of the time series of each year.

From \code{summarise_annual_time_series}, a data frame summarising the overall usage on each
day of month. Note that the \code{date_of_year} column is faked in order to put all data entries in
a leap year for comparisons.
}
\description{
Summarise and plot time series data over multiple years
}
