% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region_time_series.R
\name{plot_region_time_series}
\alias{plot_region_time_series}
\title{Create time series plot across regions for this year}
\usage{
plot_region_time_series(.data)
}
\arguments{
\item{.data}{A data frame containing the time series data.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} of the time series of this year for different regions.
}
\description{
Create time series plot across regions for this year
}
